Revert "core: move NameResolverProvider to internal"
This reverts commit 9c6ea274fe93955236c9ef4c68d02b72c9f58ccc.

The class name is important for the service loaders.
