+++ core/src/test/resources/io/grpc/NameResolverProviderTest-unavailableProvider.txt
@@ -1,0 +1,1 @@
+io.grpc.NameResolverProviderTest$UnavailableProvider
