--- core/src/test/resources/io/grpc/internal/NameResolverProviderTest-multipleProvider.txt
@@ -1,3 +1,0 @@
-io.grpc.internal.NameResolverProviderTest$Available5Provider
-io.grpc.internal.NameResolverProviderTest$Available7Provider
-io.grpc.internal.NameResolverProviderTest$Available0Provider
