--- core/src/test/resources/io/grpc/internal/NameResolverProviderTest-unavailableProvider.txt
@@ -1,1 +1,0 @@
-io.grpc.internal.NameResolverProviderTest$UnavailableProvider
